[refactor] WaypointPath: detach WaypointPath.from to mapper class with extension function
