+++ src/jvmMain/kotlin/uk/ryanwong/gmap2ics/app/models/timeline/activity/WaypointPath.kt
@@ -6,16 +6,5 @@ package uk.ryanwong.gmap2ics.app.models.timeline.activity
 
 data class WaypointPath(
     val distanceMeters: Double,
-    val roadSegmentPlaceIds: List<String>
-) {
-    companion object {
-        fun from(waypointPathDataModel: uk.ryanwong.gmap2ics.data.models.timeline.WaypointPath): WaypointPath {
-            return WaypointPath(
-                distanceMeters = waypointPathDataModel.distanceMeters ?: 0.0,
-                roadSegmentPlaceIds = waypointPathDataModel.roadSegment?.mapNotNull { roadSegment ->
-                    roadSegment.placeId
-                } ?: emptyList()
-            )
-        }
-    }
-}
+    val roadSegmentPlaceIds: List<String>,
+)
